+++ src/solidmodels/render.jl
@@ -924,11 +924,13 @@ function render!(
     _postrender!(sm, postrender_ops)
     # Get rid of redundant entities and update groups accordingly.
     # The first [1,0] call improves robustness of the next fragment significantly.
-    # Additionally, the decreasing order of dimensions is important, as OCC can error
-    # sometimes if the higher dimensional entities are not reconciled first.
     _synchronize!(sm)
     _fragment_and_map!(sm, [1, 0]) # Important!
-    _fragment_and_map!(sm, [3, 2, 1])
+    # The order of dimensions is important. There may be OCC errors or failures
+    # to map pre-fragment physical groups to post-fragment entities if this operation is reordered or
+    # broken up. For example, with [3, 2, 1], exterior boundaries could be lost from an
+    # "exterior_boundaries" physical group.
+    _fragment_and_map!(sm, [1, 2, 3])
 
     # Pass in call back function for meshing against the vertices found previously.
     gmsh.model.mesh.setSizeCallback(gmsh_meshsize)
@@ -1002,7 +1004,7 @@ end
 # excluded_physical_groups are physical groups not to be included in the fragmentation.
 function _fragment_and_map!(
     sm::SolidModel,
-    frag_dims=[3, 2, 1, 0];
+    frag_dims;
     excluded_physical_groups=PhysicalGroup[]
 )
     gmsh.model.set_current(name(sm))